[UPDATE]stripe for firebase[DELETE]pages/stripe.vue
