+++ functions/src/index.ts
@@ -31,7 +31,7 @@ const stripeCharge = (request: functions.Request, response: functions.Response) 
   const body = request.body;
   const amount = body.charge.amount;
   const currency = body.charge.currency;
-  const description = 'グランピング支払い';
+  const description = 'グランピング支払い'; // ここはあとで変えルカ
   const receipt_email = body.charge.email;
   const source = body.token.id;
   const charge = {amount, currency, description, receipt_email, source};
