+++ pages/reservation/_id.vue
@@ -6,7 +6,7 @@
 
     <v-row dense>
       <v-col :cols="10" class="ma-auto">
-        <Plan />
+        <Plan :pay-number="planPay" />
       </v-col>
 
       <v-col :cols="10" class="ma-auto">
@@ -57,7 +57,7 @@
                   >
                     <Options
                       :plan-title="optionList.title"
-                      :pay="optionList.pay"
+                      :pay="optionList.pay.toLocaleString()"
                       :texts="optionList.text"
                       :image="optionList.image"
                       :display-name="optionList.displayName"
@@ -78,16 +78,13 @@
       </v-col>
     </v-row>
 
-    <div class="text-center mt-6">
-      <v-btn color="success">
-        予約する
-      </v-btn>
-    </div>
+    <Stripe />
   </div>
 </template>
 
 <script lang="ts">
 import { Component, Vue } from 'nuxt-property-decorator'
+import Stripe from '~/components/Btn/Stripe.vue'
 import Plan from '~/components/Card/Reservation/Plan.vue'
 import Options from '~/components/Card/Reservation/Options.vue'
 
@@ -103,7 +100,7 @@ interface option {
 interface options {
   title: string
   text: [...string[]]
-  pay: string
+  pay: number
   image: string
   displayName: string
 }
@@ -111,7 +108,8 @@ interface options {
 @Component({
   components: {
     Plan,
-    Options
+    Options,
+    Stripe
   }
 })
 export default class reservation extends Vue {
@@ -119,6 +117,7 @@ export default class reservation extends Vue {
   dates: [] = []
   length: number = 0
   pageSlice: number = 0
+  payNum: number = 2000
 
   created() {
     if (window.parent.screen.width <= 420) {
@@ -132,42 +131,42 @@ export default class reservation extends Vue {
     {
       title: 'バーベキュー',
       text: ['美味しい', '楽しい'],
-      pay: '2000',
+      pay: 2000,
       image: '40',
       displayName: 'option1'
     },
     {
       title: '誕生日ケーキ',
       text: ['美味しい', '楽しい'],
-      pay: '1500',
+      pay: 1500,
       image: '10',
       displayName: 'option2'
     },
     {
       title: '国産牛肉',
       text: ['高い', 'けど、うまい！'],
-      pay: '20000',
+      pay: 20000,
       image: '16',
       displayName: 'option3'
     },
     {
       title: '誕生日ケーキ',
       text: ['美味しい', '楽しい'],
-      pay: '1500',
+      pay: 1500,
       image: '17',
       displayName: 'option4'
     },
     {
       title: '国産牛肉',
       text: ['高い', 'けど、うまい！'],
-      pay: '20000',
+      pay: 20000,
       image: '20',
       displayName: 'option5'
     },
     {
       title: 'バーベキュー',
       text: ['美味しい', '楽しい'],
-      pay: '2000',
+      pay: 2000,
       image: '30',
       displayName: 'option6'
     }
@@ -176,6 +175,10 @@ export default class reservation extends Vue {
 
   public get dateRangeText(): string {
     return this.dates.join(' ~ ')
+  }
+
+  get planPay(): number {
+    return this.$store.state.reservation.planPay
   }
 
   mounted() {
