+++ pages/reservation/complete/_id.vue
@@ -1,0 +1,9 @@
+<template>
+  <div></div>
+</template>
+
+<script lang="ts">
+import { Component, Vue } from 'nuxt-property-decorator'
+@Component
+export default class reservationComplete extends Vue {}
+</script>
